Introduce RetrieveEsiRepsonseTrait and refactor job action classes
