+++ src/Actions/Jobs/Character/CharacterInfoAction.php
@@ -2,14 +2,14 @@
 
 namespace Seatplus\Eveapi\Actions\Jobs\Character;
 
-use Seatplus\Eveapi\Actions\Eseye\RetrieveEsiDataAction;
-use Seatplus\Eveapi\Containers\EsiRequestContainer;
 use Seatplus\Eveapi\Containers\JobContainer;
 use Seatplus\Eveapi\Jobs\Alliances\AllianceInfo;
 use Seatplus\Eveapi\Models\Character\CharacterInfo;
+use Seatplus\Eveapi\Traits\RetrieveEsiResponse;
 
 class CharacterInfoAction
 {
+    use RetrieveEsiResponse;
     /**
      * @var string
      */
@@ -25,27 +25,17 @@ class CharacterInfoAction
      */
     protected $version = 'v4';
 
-    protected $retrieve_action;
-
-    public function __construct()
-    {
-        $this->retrieve_action = new RetrieveEsiDataAction();
-    }
-
     public function execute(int $character_id)
     {
-        $response = $this->retrieve_action->execute(new EsiRequestContainer([
-            'method' => $this->method,
-            'version' => $this->version,
-            'endpoint' => $this->endpoint,
-            'path_values' => [
-                'character_id' => $character_id,
-            ],
-        ]));
+        $response = $this->retrieve([
+            'character_id' => $character_id,
+        ]);
 
         if ($response->isCachedLoad()) return;
 
-        CharacterInfo::firstOrNew(['character_id' => $character_id])->fill([
+        CharacterInfo::updateOrCreate([
+            'character_id' => $character_id
+        ], [
             'name'            => $response->name,
             'description'     => $response->optional('description'),
             'corporation_id'  => $response->corporation_id,
@@ -58,7 +48,7 @@ class CharacterInfoAction
             'security_status' => $response->optional('security_status'),
             'faction_id'      => $response->optional('faction_id'),
             'title' => $response->optional('title'),
-        ])->save();
+        ]);
 
         if (! empty($response->optional('alliance_id')))
         {
